Refactor last traces of "Universal" nomenclature
