+++ lib/api/Cargo.toml
@@ -105,7 +105,7 @@ default-llvm = ["default-compiler", "llvm"]
 # - Engines.
 engine = ["sys"]
 universal = [
-    "engine", "wasmer-compiler/universal_engine"
+    "engine", "wasmer-compiler/engine_compilation"
 ]
 default-engine = []
 default-universal = [
