+++ lib/cli-compiler/Cargo.toml
@@ -18,7 +18,7 @@ path = "src/bin/wasmer_compiler.rs"
 doc = false
 
 [dependencies]
-wasmer-compiler = { version = "=2.3.0", path = "../compiler", features = ["universal_engine"] }
+wasmer-compiler = { version = "=2.3.0", path = "../compiler", features = ["engine_compilation"] }
 wasmer-types = { version = "=2.3.0", path = "../types" }
 atty = "0.2"
 colored = "2.0"
