+++ lib/cli/Cargo.toml
@@ -26,7 +26,7 @@ required-features = ["headless"]
 
 [dependencies]
 wasmer = { version = "=2.3.0", path = "../api", default-features = false }
-wasmer-compiler = { version = "=2.3.0", path = "../compiler", features = ["universal_engine", ] }
+wasmer-compiler = { version = "=2.3.0", path = "../compiler", features = ["engine_compilation", ] }
 wasmer-compiler-cranelift = { version = "=2.3.0", path = "../compiler-cranelift", optional = true }
 wasmer-compiler-singlepass = { version = "=2.3.0", path = "../compiler-singlepass", optional = true }
 wasmer-compiler-llvm = { version = "=2.3.0", path = "../compiler-llvm", optional = true }
