+++ lib/compiler/src/artifact_builders/engine_builder.rs
