+++ lib/compiler/src/artifact_builders/trampoline.rs
