+++ lib/compiler/src/engine/code_memory.rs
